+++ packages/create-openfort/src/utils/renderVersionWarning.ts
@@ -11,15 +11,15 @@ export const renderVersionWarning = (npmVersion: string) => {
   //   console.log("npm", npmVersion);
 
   if (currentVersion.includes("beta")) {
-    logger.warn("  You are using a beta version of create-openfort-app.");
+    logger.warn("  You are using a beta version of create-openfort.");
     logger.warn("  Please report any bugs you encounter.");
   } else if (currentVersion.includes("next")) {
     logger.warn(
-      "  You are running create-openfort-app with the @next tag which is no longer maintained.",
+      "  You are running create-openfort with the @next tag which is no longer maintained.",
     );
     logger.warn("  Please run the CLI with @latest instead.");
   } else if (currentVersion !== npmVersion) {
-    logger.warn("  You are using an outdated version of create-openfort-app.");
+    logger.warn("  You are using an outdated version of create-openfort.");
     logger.warn(
       "  Your version:",
       `${currentVersion}.`,
@@ -45,7 +45,7 @@ function checkForLatestVersion(): Promise<string> {
   return new Promise((resolve, reject) => {
     https
       .get(
-        "https://registry.npmjs.org/-/package/create-openfort-app/dist-tags",
+        "https://registry.npmjs.org/-/package/create-openfort/dist-tags",
         (res) => {
           if (res.statusCode === 200) {
             let body = "";
@@ -73,7 +73,7 @@ export const getNpmVersion = () =>
   // `fetch` to the registry is faster than `npm view` so we try that first
   checkForLatestVersion().catch(() => {
     try {
-      return execSync("npm view create-openfort-app version").toString().trim();
+      return execSync("npm view create-openfort version").toString().trim();
     } catch {
       return null;
     }